Fixed comments. Moved updateTextElement to dom helpers
